refactor: 당첨 로또를 구하는 기능 분리
